requirements.txt
alembic==1.9.4
aniso8601==9.0.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
exceptiongroup==1.1.0
fastapi==0.92.0
graphene==3.2.1
graphene-sqlalchemy==2.1.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mongoengine==0.26.0
motor==3.1.1
mypy-extensions==1.0.0
odmantic==0.9.2
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
psycopg2-binary==2.9.3
pydantic==1.10.5
pymongo==4.3.3
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
singledispatch==4.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.25.0
starlette-graphene3==0.6.0
strawberry-graphql==0.156.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
